# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/marcomarcogd/MRSS/issues"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.4.2"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
...
        "/articles": {
            "get": {
                "description": "Retrieve articles with optional filtering by feed, category, status, and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get articles with filtering",
                "parameters": [
                    {
                        "enum": [
                            "all",
                            "unread",
                            "favorite",
                            "read_later"
                        ],
                        "type": "string",
                        "description": "Filter: 'all', 'unread', 'favorite', 'read_later'",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Filter by feed ID",
                        "name": "feed_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by category name",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 500,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Items per page (default: 50, max: 500)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Article"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/articles/export/obsidian": {
            "post": {
                "description": "Export an article to Obsidian vault as a Markdown file (requires obsidian_enabled and obsidian_vault_path settings)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Export article to Obsidian",
                "parameters": [
                    {
                        "description": "Article export request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/article.ExportToObsidianRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export result (success, file_path, message)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (Obsidian not configured or invalid article ID)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/opml/export": {
            "get": {
                "description": "Export all feeds as an OPML file (server mode - direct download)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/xml"
                ],
                "tags": [
                    "opml"
                ],
                "summary": "Export OPML file",
                "responses": {
                    "200": {
                        "description": "OPML XML file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/translate/article": {
            "post": {
                "description": "Translate an article's title to the target language (uses AI or Google based on settings)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "translation"
                ],
                "summary": "Translate article title",
                "parameters": [
                    {
                        "description": "Translation request (article_id, title, target_language)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation result (translated_title, limit_reached)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (missing required fields)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "article.ExportToObsidianRequest": {
            "type": "object",
            "properties": {
                "article_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Article": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "author": {
                    "description": "Article author",
                    "type": "string"
                },
                "feed_id": {
                    "type": "integer"
                },
                "feed_title": {
                    "description": "Joined field",
                    "type": "string"
                },
                "freshrss_item_id": {
                    "description": "FreshRSS/Google Reader item ID for API operations",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "is_hidden": {
                    "type": "boolean"
                },
                "is_read": {
                    "type": "boolean"
                },
                "is_read_later": {
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "summary": {
                    "description": "Cached AI-generated summary",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "translated_title": {
                    "type": "string"
                },
                "unique_id": {
                    "description": "Unique identifier for deduplication (title+feed_id+published_date)",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "video_url": {
                    "description": "YouTube video URL for embedded player",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}